Support compilerfolder and online environments (#3607)
When specifying BcAuthContext and Environment to Run-AlPipeline,
Run-AlPipeline would always create a filesonly container, disallowing
running on Linux.
This PR fixes this plus some bugs found as a result of this.

- Replace all occurrences of
[System.Runtime.InteropServices.Marshal]::PtrToStringAuto with
[System.Runtime.InteropServices.Marshal]::PtrToStringBSTR as the Auto
function doesn't always do what's expected under Linux (We do not use
the Auto function in AL-Go)
- Ensure correct casing of Newtonsoft.Json.dll for Linux (also not a
problem in AL-Go)
- Always add extensionId (when specified) to Properties section in test
results xml
- Also added two new overrides (PipelineInitialize and PipelineFinalize)
requested by COSMO Consult.
- If environment is specified as a Web Client URL, and BcAuthContext
contains username/password in Run-AlPipeline, then tests will run
against this environment. PublishBcContainerApp and
ImportTestToolkitToBcContainer needs to be overridden for this to work
with full pipeline.
- Add parameter CompilerFolder to Run-TestsInBcContainer and
Import-TestToolkitToBcContainer for running tests using CompilerFolder
bits from the host
- Including caching of appinfos in CompilerFolder cache (to save time
when caching on GitHub Actions)

Running Build AND Test under Linux (using CompilerFolder), using an
online environment as "Service Tier" can be seen here:

https://github.com/BusinessCentralDemos/bingmaps.pte/actions/runs/10313615507

Build and test here takes approx. 3 minutes.

This functionality is needed by COSMO to enable using their Docker Swarm
for running tests in AL-Go.

COSMO is aware that AL-Go moves away from using BcContainerHelper and
will subsequently have to change their integration when this has
happened.

---------

+++ AppHandling/Run-TestsInNavContainer.ps1
@@ -87,6 +87,7 @@
 function Run-TestsInBcContainer {
     Param (
         [string] $containerName = $bcContainerHelperConfig.defaultContainerName,
+        [string] $compilerFolder = '',
         [Parameter(Mandatory=$false)]
         [string] $tenant = "default",
         [Parameter(Mandatory=$false)]
@@ -141,23 +142,60 @@ function Run-TestsInBcContainer {
 
 $telemetryScope = InitTelemetryScope -name $MyInvocation.InvocationName -parameterValues $PSBoundParameters -includeParameters @()
 try {
-    
-    $customConfig = Get-BcContainerServerConfiguration -ContainerName $containerName
-    $navversion = Get-BcContainerNavversion -containerOrImageName $containerName
-    $version = [System.Version]($navversion.split('-')[0])
+
+    if ($containerName) {
+        Write-Host "Using Container"
+        $customConfig = Get-BcContainerServerConfiguration -ContainerName $containerName
+        $navversion = Get-BcContainerNavversion -containerOrImageName $containerName
+        $version = [System.Version]($navversion.split('-')[0])
+        $PsTestToolFolder = Join-Path $bcContainerHelperConfig.hostHelperFolder "Extensions\$containerName\PsTestTool"
+
+    }
+    elseif ($compilerFolder) {
+        Write-Host "Using CompilerFolder"
+        $customConfig = $null
+        $symbolsFolder = Join-Path $compilerFolder "symbols"
+        $baseAppInfo = Get-AppJsonFromAppFile -appFile (Get-ChildItem -Path $symbolsFolder -Filter 'Microsoft_Base Application_*.*.*.*.app').FullName
+        $version = [Version]$baseAppInfo.version
+        $PsTestToolFolder = Join-Path ([System.IO.Path]::GetTempPath()) "$([Guid]::NewGuid().ToString())"
+        New-Item $PsTestToolFolder -ItemType Directory | Out-Null
+        $testDlls = Join-Path $compilerFolder "dlls/Test Assemblies/*.dll"
+        Copy-Item $testDlls -Destination $PsTestToolFolder -Force
+        Copy-Item -Path (Join-Path $PSScriptRoot "PsTestFunctions.ps1") -Destination $PsTestToolFolder -Force
+        Copy-Item -Path (Join-Path $PSScriptRoot "ClientContext.ps1") -Destination $PsTestToolFolder -Force
+    }
+    else {
+        throw "You must specify either containerName or compilerFolder"
+    }
 
     if ($bcAuthContext -and $environment) {
-        $response = Invoke-RestMethod -Method Get -Uri "$($bcContainerHelperConfig.baseUrl.TrimEnd('/'))/$($bcAuthContext.tenantID)/$environment/deployment/url"
-        if($response.status -ne 'Ready') {
-            throw "environment not ready, status is $($response.status)"
+        if ($environment -like 'https://*') {
+            $useUrl = $environment
+            if ($bcAuthContext.ContainsKey('Username') -and $bcAuthContext.ContainsKey('Password')) {
+                $credential = New-Object System.Management.Automation.PSCredential -ArgumentList $bcAuthContext.Username, $bcAuthContext.Password
+                $clientServicesCredentialType = "NavUserPassword"
+            }
+            if ($bcAuthContext.ContainsKey('ClientServicesCredentialType')) {
+                $clientServicesCredentialType = $bcAuthContext.ClientServicesCredentialType
+            }
+            $testPage = 130455
         }
-        $useUrl = $response.data.Split('?')[0]
-        $tenant = ($response.data.Split('?')[1]).Split('=')[1]
-
-        if ($testPage) {
-            throw "You cannot specify testPage when running tests in an Online tenant"
+        else {
+            $response = Invoke-RestMethod -Method Get -Uri "$($bcContainerHelperConfig.baseUrl.TrimEnd('/'))/$($bcAuthContext.tenantID)/$environment/deployment/url"
+            if($response.status -ne 'Ready') {
+                throw "environment not ready, status is $($response.status)"
+            }
+            $useUrl = $response.data
+            if ($testPage) {
+                throw "You cannot specify testPage when running tests in an Online tenant"
+            }
+            $testPage = 130455
         }
-        $testPage = 130455
+        $uri = [Uri]::new($useUrl)
+        $useUrl = $useUrl.Split('?')[0]
+        $dict = [System.Web.HttpUtility]::ParseQueryString($uri.Query)
+        if ($dict['tenant']) { $tenant = $dict['tenant'] }
+        if ($dict['testpage']) { $testpage = [int]$dict['testpage'] }
     }
     else {
         $clientServicesCredentialType = $customConfig.ClientServicesCredentialType
@@ -220,13 +258,12 @@ try {
         } -argumentList $interactionTimeout.ToString()
     }
 
-    if ($bcAuthContext) {
+    if ($bcAuthContext -and ($environment -notlike 'https://*')) {
         $bcAuthContext = Renew-BcAuthContext $bcAuthContext
         $accessToken = $bcAuthContext.accessToken
         $credential = New-Object pscredential -ArgumentList $bcAuthContext.upn, (ConvertTo-SecureString -String $accessToken -AsPlainText -Force)
     }
 
-    $PsTestToolFolder = Join-Path $bcContainerHelperConfig.hostHelperFolder "Extensions\$containerName\PsTestTool"
     $PsTestFunctionsPath = Join-Path $PsTestToolFolder "PsTestFunctions.ps1"
     $ClientContextPath = Join-Path $PsTestToolFolder "ClientContext.ps1"
     $fobfile = Join-Path $PsTestToolFolder "PSTestToolPage.fob"
@@ -273,24 +310,30 @@ try {
         try
         {
             if ($connectFromHost) {
-                $newtonSoftDllPath = Join-Path $PsTestToolFolder "NewtonSoft.json.dll"
+                if ($PSVersionTable.PSVersion.Major -lt 7) {
+                    throw "Using ConnectFromHost requires PowerShell 7"
+                }
+                $newtonSoftDllPath = Join-Path $PsTestToolFolder "Newtonsoft.Json.dll"
                 $clientDllPath = Join-Path $PsTestToolFolder "Microsoft.Dynamics.Framework.UI.Client.dll"
-    
-                Invoke-ScriptInBcContainer -containerName $containerName { Param([string] $myNewtonSoftDllPath, [string] $myClientDllPath)
-                
-                    if (!(Test-Path $myNewtonSoftDllPath)) {
-                        $newtonSoftDllPath = "C:\Program Files\Microsoft Dynamics NAV\*\Service\Management\NewtonSoft.json.dll"
-                        if (!(Test-Path $newtonSoftDllPath)) {
-                            $newtonSoftDllPath = "C:\Program Files\Microsoft Dynamics NAV\*\Service\NewtonSoft.json.dll"
-                        }
-                        $newtonSoftDllPath = (Get-Item $newtonSoftDllPath).FullName
-                        Copy-Item -Path $newtonSoftDllPath -Destination $myNewtonSoftDllPath
+                if ($containerName) {
+                    if (!((Test-Path $newtonSoftDllPath) -and (Test-Path $clientDllPath))) {
+                        Invoke-ScriptInBcContainer -containerName $containerName { Param([string] $myNewtonSoftDllPath, [string] $myClientDllPath)
+                    
+                            if (!(Test-Path $myNewtonSoftDllPath)) {
+                                $newtonSoftDllPath = "C:\Program Files\Microsoft Dynamics NAV\*\Service\Management\Newtonsoft.Json.dll"
+                                if (!(Test-Path $newtonSoftDllPath)) {
+                                    $newtonSoftDllPath = "C:\Program Files\Microsoft Dynamics NAV\*\Service\Newtonsoft.Json.dll"
+                                }
+                                $newtonSoftDllPath = (Get-Item $newtonSoftDllPath).FullName
+                                Copy-Item -Path $newtonSoftDllPath -Destination $myNewtonSoftDllPath
+                            }
+                            $clientDllPath = "C:\Test Assemblies\Microsoft.Dynamics.Framework.UI.Client.dll"
+                            if (!(Test-Path $myClientDllPath)) {
+                                Copy-Item -Path $clientDllPath -Destination $myClientDllPath
+                            }
+                        } -argumentList $newtonSoftDllPath, $clientDllPath
                     }
-                    $clientDllPath = "C:\Test Assemblies\Microsoft.Dynamics.Framework.UI.Client.dll"
-                    if (!(Test-Path $myClientDllPath)) {
-                        Copy-Item -Path $clientDllPath -Destination $myClientDllPath
-                    }
-                } -argumentList $newtonSoftDllPath, $clientDllPath
+                }
     
                 if ($useUrl) {
                     $publicWebBaseUrl = $useUrl.TrimEnd('/')
@@ -384,11 +427,11 @@ try {
                     }
                 }
 
-                $result = Invoke-ScriptInBcContainer -containerName $containerName -usePwsh $false -scriptBlock { Param([string] $tenant, [string] $companyName, [string] $profile, [pscredential] $credential, [string] $accessToken, [string] $testSuite, [string] $testGroup, [string] $testCodeunit, [string] $testCodeunitRange, [string] $testFunction, [string] $PsTestFunctionsPath, [string] $ClientContextPath, [string] $XUnitResultFileName, [bool] $AppendToXUnitResultFile, [string] $JUnitResultFileName, [bool] $AppendToJUnitResultFile, [bool] $ReRun, [string] $AzureDevOps, [string] $GitHubActions, [bool] $detailed, [timespan] $interactionTimeout, $testPage, $version, $culture, $timezone, $debugMode, $usePublicWebBaseUrl, $useUrl, $extensionId, $testRunnerCodeunitId, $disabledtests, $renewClientContextBetweenTests)
+                $result = Invoke-ScriptInBcContainer -containerName $containerName -usePwsh $false -scriptBlock { Param([string] $tenant, [string] $companyName, [string] $profile, [System.Management.Automation.PSCredential] $credential, [string] $accessToken, [string] $testSuite, [string] $testGroup, [string] $testCodeunit, [string] $testCodeunitRange, [string] $testFunction, [string] $PsTestFunctionsPath, [string] $ClientContextPath, [string] $XUnitResultFileName, [bool] $AppendToXUnitResultFile, [string] $JUnitResultFileName, [bool] $AppendToJUnitResultFile, [bool] $ReRun, [string] $AzureDevOps, [string] $GitHubActions, [bool] $detailed, [timespan] $interactionTimeout, $testPage, $version, $culture, $timezone, $debugMode, $usePublicWebBaseUrl, $useUrl, $extensionId, $testRunnerCodeunitId, $disabledtests, $renewClientContextBetweenTests)
     
-                    $newtonSoftDllPath = "C:\Program Files\Microsoft Dynamics NAV\*\Service\Management\NewtonSoft.json.dll"
+                    $newtonSoftDllPath = "C:\Program Files\Microsoft Dynamics NAV\*\Service\Management\Newtonsoft.Json.dll"
                     if (!(Test-Path $newtonSoftDllPath)) {
-                        $newtonSoftDllPath = "C:\Program Files\Microsoft Dynamics NAV\*\Service\NewtonSoft.json.dll"
+                        $newtonSoftDllPath = "C:\Program Files\Microsoft Dynamics NAV\*\Service\Newtonsoft.Json.dll"
                     }
                     $newtonSoftDllPath = (Get-Item $newtonSoftDllPath).FullName
                     $clientDllPath = "C:\Test Assemblies\Microsoft.Dynamics.Framework.UI.Client.dll"
